refactor: split claude adapter into claude/ folder
