+++ apps/backend/src/agent-core/llm-api/llm-api.ts
@@ -1,4 +1,4 @@
-import {streamClaude} from './claude-adapter.js';
+import {streamClaude} from './claude/index.js';
 import {streamOpenAI} from './openai-adapter.js';
 import {streamOpenAIResponses} from './openai-responses-adapter.js';
 import type {LlmCompletionOptions, LlmEventStream} from './types.js';
